Moved Dialogs to a dedicated Class. Dialog DOES NOT WORK yet, as Dave & I spoke with StevenIc who has changes coming. As the exemplar for Dialogs is finalized, this code will be updated.
