+++ samples-final/8.AspNetCore-LUIS-Bot/MyLuisBot.cs
@@ -9,7 +9,6 @@ using System.Threading.Tasks;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Ai.Luis;
 using Microsoft.Bot.Builder.Dialogs;
-using Microsoft.Bot.Builder.Dialogs.Choices;
 using Microsoft.Bot.Schema;
 
 namespace AspNetCore_LUIS_Bot
@@ -18,96 +17,26 @@ namespace AspNetCore_LUIS_Bot
     {
         private const double LUIS_INTENT_THRESHOLD = 0.2d;
 
-        private readonly DialogSet _dialogs;        
         private readonly MyBotAccessors _stateAccessors;
-        private readonly LuisRecognizer _luisRecognizer;
+        private readonly LuisRecognizer _luisRecognizer;        
+        private readonly DialogSet _dialogSet;
 
         public MyLuisBot(MyBotAccessors accessors, LuisRecognizer recognizer)
-        {            
+        {
             _stateAccessors = accessors ?? throw new ArgumentNullException(nameof(accessors));
             _luisRecognizer = recognizer ?? throw new ArgumentNullException(nameof(recognizer));
 
-            _dialogs = new DialogSet();
-            _dialogs.Add("None", new WaterfallStep[] { DefaultDialog });
-            _dialogs.Add("Calendar_Add", new WaterfallStep[] { AskReminderTitle, SaveReminder });
-            _dialogs.Add("Calendar_Find", new WaterfallStep[] { ShowReminders, ConfirmShow });
-            _dialogs.Add("TitlePrompt", new TextPrompt(TitleValidator));
-            _dialogs.Add("ShowReminderPrompt", new ChoicePrompt(Microsoft.Recognizers.Text.Culture.English));
-        }
+            _dialogSet = new DialogSet();
 
-        private async Task TitleValidator(ITurnContext context, TextResult result)
-        {
-            if (string.IsNullOrWhiteSpace(result.Value) || result.Value.Length < 3)
-            {
-                result.Status = PromptStatus.NotRecognized;
-                await context.SendActivityAsync("Title should be at least 3 characters long.");
-            }
-        }
-
-        private Task DefaultDialog(DialogContext dialogContext, object args, SkipStepFunction next)
-        {
-            return dialogContext.Context.SendActivityAsync("Hi! I'm a simple reminder bot. I can add reminders and show them.");
-        }
-
-        private async Task AskReminderTitle(DialogContext dialogContext, object args, SkipStepFunction next)
-        {
-            var dialogState = await dialogContext.Context.GetStateAsync<Dictionary<string, object>>(_stateAccessors.UserDialogState);
-            if (dialogState.ContainsKey("Title"))
-            {
-                await dialogContext.ContinueAsync();
-            }
-            else
-            {
-                await dialogContext.PromptAsync("TitlePrompt", "What would you like to call your reminder?");
-            }
-        }
-
-        private async Task SaveReminder(DialogContext dialogContext, object args, SkipStepFunction next)
-        {
-            var title = "";
-            if (args is TextResult textResult)
-            {
-                title = textResult.Value;
-            }
-
-            // Update existing property
-
-            var titles = await dialogContext.Context.GetStateAsync<List<Reminder>>(_stateAccessors.Reminders);
-
-            titles.Add(new Reminder() { Title = title } );
-            await dialogContext.Context.SetStateAsync<List<Reminder>>(_stateAccessors.Reminders, titles);
-
-
-            await dialogContext.Context.SendActivityAsync($"Your reminder named '{title}' is set.");
-            await dialogContext.EndAsync();
-        }
-
-        private async Task ShowReminders(DialogContext dialogContext, object args, SkipStepFunction next)
-        {
-            var titles = await dialogContext.Context.GetStateAsync<List<Reminder>>(_stateAccessors.Reminders);
-
-            var choices = titles.Select(x => new Choice() { Value = x.Title.Length < 15 ? x.Title : x.Title.Substring(0, 15) + "..." }).ToList();
-            await dialogContext.PromptAsync("ShowReminderPrompt", "Select the reminder to show: ", new ChoicePromptOptions() { Choices = choices });
-        }
-
-        private async Task ConfirmShow(DialogContext dialogContext, object args, SkipStepFunction next)
-        {
-
-            if (args is ChoiceResult choice)
-            {
-                var reminders = await dialogContext.Context.GetStateAsync<List<Reminder>>(_stateAccessors.Reminders);
-                var reminder = ((List<Reminder>)reminders)[choice.Value.Index];
-
-                await dialogContext.Context.SendActivityAsync($"Reminder: {reminder.Title}");
-            }
-            await dialogContext.EndAsync();
+            var reminderDialog = new ReminderDialog(_stateAccessors);            
+            _dialogSet.Add("AlarmDialog", reminderDialog); 
         }
 
         public async Task OnTurnAsync(ITurnContext turnContext, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (turnContext == null)
             {
-                throw new System.ArgumentNullException(nameof(turnContext));
+                throw new ArgumentNullException(nameof(turnContext));
             }
 
             if (turnContext.Activity.Type == ActivityTypes.ConversationUpdate && turnContext.Activity.MembersAdded.FirstOrDefault()?.Id == turnContext.Activity.Recipient.Id)
@@ -118,7 +47,7 @@ namespace AspNetCore_LUIS_Bot
             {
                 var dialogState = await turnContext.GetStateAsync<Dictionary<string, object>>(_stateAccessors.UserDialogState);
 
-                var dialogContext = _dialogs.CreateContext(turnContext, dialogState);
+                var dialogContext = _dialogSet.CreateContext(turnContext, dialogState);
 
                 var utterance = turnContext.Activity.Text.ToLowerInvariant();
                 if (utterance == "cancel")
@@ -140,10 +69,12 @@ namespace AspNetCore_LUIS_Bot
 
                     if (!turnContext.Responded)
                     {
+                        //TODO: This code will be updated with new Dialogs as the revised models land
+                        // and the LUIS + Dialog examplar is formalized. 
                         var luisResult = await _luisRecognizer.RecognizeAsync(turnContext, CancellationToken.None);
                         var (intent, score) = luisResult.GetTopScoringIntent();
                         var intentResult = score > LUIS_INTENT_THRESHOLD ? intent : "None";
-                        await dialogContext.BeginAsync(intent);
+                        await dialogContext.BeginAsync("AlarmDialog");
                     }
                 }
             }
